# Improve performance 3-fold