--- a/setup.py
+++ b/setup.py
@@ -6,7 +6,7 @@ os.chdir(os.path.normpath(os.path.join(os.path.abspath(__file__), os.pardir)))
 
 setup(
     name='dataclasses_json',
-    version='1.0.7',
+    version='1.0.8',
     packages=['dataclasses_json'],
     include_package_data=True,
     description="Allows for easy dataclass parsing from/to JSON",