@cakemanny 2019-11-27
NLiven uses a fair amount of CPU, so I'd be interested to see if there becomes a noticeable change
![image](https://user-images.githubusercontent.com/4025945/69753589-dd39e600-114b-11ea-96f2-ce5f368e3c4c.png)