@exaroth 2019-11-27
looks good, we do versioning of this lib in requirements so i think this can be merged there, however before updating all adapters with it id probably push it for some low profile one and leave it for a week (tourcms/prioticket for example) as im not sure if it it wont have any unpredictable consequences to the adapters. Btw original version of dataclasses json still has that deepcopy in there https://github.com/lidatong/dataclasses-json/blob/master/dataclasses_json/core.py#L294