@exaroth 2019-11-27
Also 2 things i would consider to improve this lib:

1. As you mentioned, remove that shitty recursive calls
2. Add faster json serializer